rename and namespace things
