+++ src/utils.js
@@ -1,7 +1,7 @@
 // @flow
 import { memoize } from 'lodash'
 import type { SpacingProps, Noop, SpacingValues } from './types'
-import type { SpacingAliases } from './spacingAliasesProvider'
+import type { SpacingAliases } from './configProvider'
 
 const isProd = process.env.NODE_ENV === 'production'
 const isObjPropNumber = keys => key => typeof keys[key] === 'number'
